reorganize for firebase
