+++ .firebaserc
@@ -1,0 +1,5 @@
+{
+  "projects": {
+    "default": "team-04-comp-2930"
+  }
+}
