--- README.md
@@ -1,2 +1,0 @@
-# EmissionsPlanner
-COMP 2930 Project Emissions Planner
