--- css/addVehicle.css
@@ -1,307 +1,0 @@
-/* Page */
-
-html {
-    box-sizing: border-box;
-    margin: 0;
-}
-
-body {
-    margin: 0;
- 
-}
-
-*, 
-*:before, 
-*:after {
-    box-sizing: inherit;
-}
-
-
-.tabs {
-    margin:0 auto 0 auto;
-    width: 100%;
-    background-color: rgba(210, 235, 191, 1);
-
-}
-
-.tabs-content {
-    color: rgb(59, 59, 59);
-}
-
-/* Tabs */
-
-.tabs-nav {
-    overflow: hidden;
-}
-
-.tabs-nav-items {
-    margin: 0;
-    padding: 0;
-    padding-top: 10px;
-    padding-left: 10px;
-    list-style: none;
-    border-bottom: 3px solid #ffffff;
-}
-
-.tabs-nav-items:before,
-.tabs-nav-items:after {
-    content: " ";
-    display: table;
-}
-
-.tabs-nav-items:after {
-    clear: both;
-}
-
-.tabs-nav-item {
-    float: left;
-    position: relative;
-    width: 25%;
-    padding: 0;
-    cursor: pointer;
-}
-
-.tabs-nav-item:last-of-type {
-    margin-right: 0;
-}
-
-.tabs-nav-item::before {
-   
-    content: '';
-    position: absolute;
-    display: block;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: #ffffff;
-    border-radius: 3px 3px 0 0;
-    -webkit-transform: translate(0, 100%);
-    transform: translate(0, 100%);
-    transition-property: transform;
-}
-
-.tabs-nav-btn {
-
-    position: relative;
-    display: block;
-    outline: 0 none;
-    text-decoration: none;
-    padding: 13px 0;
-    color: rgb(59, 59, 59);
-    font-family: segoe ui,Arial,sans-serif;
-    font-size: 2.5vw;
-    font-weight: bold;
-    letter-spacing: 1.5px;
-    text-align: center;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    transition-property: color;
-}
-
-.tabs-nav-item::before,
-.tabs-nav-btn {
-
-    transition-duration: 0.2s;
-    transition-timing-function: cubic-bezier(0.7,0,0.3,1);
-}
-
-.tabs-content {
-    border-top: 0;
-    background-color: rgb(255, 255, 255);
-
-}
-
-#first{
-    background-color: rgb(255, 255, 255);
-}
-
-.tabs-content-item {
-    display: none;
-    padding: 20px 0; 
-}
-
-.tabs-nav-item:hover .tabs-nav-btn {
-    color: #ffffff;
-}
-
-.active-tab-nav-item,
-.active-tab-nav-item:hover {
-    cursor: default;
-}
-
-.active-tab-nav-item:hover .tabs-nav-btn {
-    color: rgb(59, 59, 59);
-}
-
-.active-tab-nav-item::before {
-    -webkit-transform: translate(0, 0);
-    transform: translate(0, 0);
-}
-
-.active-tab-nav-item .tabs-nav-btn {
-    cursor: default;
-    color: rgb(59, 59, 59);
-}
-
-.active-tab-content-item {
-    display: block;
-}
-
-#section-1 {
-
-    padding-left:30px;
-    padding-right:30px;
-
-}
-
-#year-label {
-    margin-top: 10px;
-    background-color: rgba(210, 235, 191, 1);
-    padding: 8px;
-    
-
-}
-
-#year-label2 {
-    margin-top: 30px;
-    background-color: rgba(210, 235, 191, 1);
-    padding: 8px;
-    
-
-}
-
-.year-box {
-    width: 100%;
-    height: 35px;
-    background-color: rgb(236, 233, 233);
-    border: none;
-    border-bottom: 2px solid rgb(143, 143, 143);
-    
-
-}
-
-.button {
-    display:block;
-
-    border-radius: 4px;
-    background-color: rgba(210, 235, 191, 1);
-    border: none;
-    color: rgb(0, 0, 0);
-    text-align: center;
-    font-size: 16px;
-    padding: 10px;
-    width: 95px;
-    transition: all 0.5s;
-    cursor: pointer;
-    position: absolute;
-    margin: 30px auto;
-    justify-content: center;
-
-  left: 0;
-
-  right: 0;
-  }
-  
-  .button span {
-    cursor: pointer;
-    display: inline-block;
-    position: relative;
-    transition: 0.5s;
-  }
-  
-  .button span:after {
-    content: '+';
-    position: absolute;
-    opacity: 0;
-    top: 0;
-    right: -20px;
-    transition: 0.5s;
-  }
-  
-  .button:hover span {
-    padding-right: 25px;
-  }
-  
-  .button:hover span:after {
-    opacity: 1;
-    right: 0;
-  }
-
-  .carName {
-    padding:   10px;
-    font-size: 22px;
-    margin:0 auto 0 auto;
-    width: 100%;
-    background-color: rgba(210, 235, 191, 1);
-  }
-
-  .gray{
-      
-      padding-bottom: 10px;
-      padding-left: 10px;
-    background-color: rgb(236, 233, 233); 
-    
-  }
-
-  .not{
-      margin: 20px;
-      margin-bottom: 20px;
-  }
-
-  .button2{
-    float: right;
-    margin-right:20px;
-    background-color: rgb(209, 40, 28); /* Green */
-  border: none;
-  color: white;
-  text-align: center;
-  text-decoration: none;
-  display: inline-block;
-  font-size: 16px;
-  margin: 4px 2px;
-  cursor: pointer;
-}
-
-.button2:hover {
-    box-shadow: 0 12px 16px 0 rgba(0,0,0,0.24),0 17px 50px 0 rgba(0,0,0,0.19);
-  }
-
-  footer {
-    
-    width: 100%;
-    height: 37px;
-    padding-bottom: 20px;
-    padding-top: 5px;
-    padding-right: 20px;
-    margin-top: 20px;
-    background-color: rgba(210, 235, 191, 1);
-    position:fixed;
-    bottom: 0;
-
-  }
-  
-  h2 {
-    float: left;
-    height: 12px;
-    line-height: 12px;
-  }
-  
-  .footer-nav {
-    text-align: right;
-    margin: 0px;
-    padding: 0px;
-  }
-  
-  li {
-    display: inline-block;
-    margin-right: 20px;
-  }
-  
-  .top, .footer-nav li:last-child {
-      margin: 0px;
-    padding: 0px;
-    margin-right: 0;
-  }
-  
-
