--- css/login.css
@@ -1,149 +1,0 @@
-html,
-body {
-    height: 100%;
-
-}
-
-body {
-    display: -ms-flexbox;
-    display: flex;
-    -ms-flex-align: center;
-
-    padding-top: 40px;
-    padding-bottom: 40px;
-}
-
-.reg {
-    display: none;
-}
-
-.form-signin {
-    width: 100%;
-    max-width: 330px;
-    padding: 15px;
-    margin: auto;
-}
-
-.form-signin .checkbox {
-    font-weight: 400;
-}
-
-.form-signin .form-control {
-    position: relative;
-    box-sizing: border-box;
-    height: auto;
-    padding: 10px;
-    font-size: 16px;
-}
-
-.form-signin .form-control:focus {
-    z-index: 2;
-}
-
-.form-signin input[type="email"] {
-    margin-bottom: -1px;
-    border-bottom-right-radius: 0;
-    border-bottom-left-radius: 0;
-}
-
-.form-signin input[type="password"] {
-    margin-bottom: 10px;
-    border-top-left-radius: 0;
-    border-top-right-radius: 0;
-}
-
-#signIn {
-    background-color: rgba(210, 235, 191, 1);
-    border-color: rgba(210, 235, 191, 1);
-}
-
-#register {
-    background-color: rgba(210, 235, 191, 1);;
-    border-color: rgba(210, 235, 191, 1);
-}
-
-#signUpbtn {
-    background-color: rgba(210, 235, 191, 1);
-    border-color: rgba(210, 235, 191, 1);
-}
-
-* {
-    box-sizing: border-box;
-  }
-  
-  html, body {
-    height: 100%;
-  }
-  
-  body {
-    margin: 0;
-
-  }
-  
-  ul {
-    list-style: none;
-    padding-left: 0;
-  }
-  
-  a {
-    text-decoration: none;
-    color: white;
-    text-transform: uppercase;
-    line-height: 1.38em;
-  }
-  
-  footer {
-    position: absolute;
-    bottom: 0;
-    width: 100%;
-    height: 37px;
-    padding-bottom: 20px;
-    padding-top: 5px;
-    padding-right: 20px;
-    margin: 0;
-    background-color: rgba(210, 235, 191, 1);
-  }
-  
-  h2 {
-    float: left;
-    height: 12px;
-    line-height: 12px;
-  }
-  
-  .footer-nav {
-    text-align: right;
-    margin: 0px;
-    padding: 0px;
-  }
-  
-  li {
-    display: inline-block;
-    margin-right: 20px;
-  }
-  
-  .top, .footer-nav li:last-child {
-      margin: 0px;
-    padding: 0px;
-    margin-right: 0;
-  }
-  
-  .full {
-
-
-  }
-
-  #header{
-    position: absolute;
-    top: 0;
-    width: 100%;
-  }
-  
-  
-  @media (max-width: 768px) {
-    .top {
-
-    }
-    .full {
-        font-size: 16px;
-    }
-  }
