--- html/home.html
@@ -1,181 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-
-  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
-    integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous" />
-
-  <title>Document</title>
-  <style>
-    body {
-      min-width: 320px;
-      height: 100vh;
-    }
-
-    main {
-      background-image: url("../images/green_backimage.png");
-      background-repeat: no-repeat;
-      background-size: contain;
-    }
-
-    .btn-primary {
-      background-color: #85D139 !important
-    }
-
-    a.dropdown-item:hover {
-      background-color: #85D139;
-      color: white;
-    }
-
-    div#planTrip {
-      max-width: 570px;
-    }
-
-    div#settingIcon button {
-      border-color: rgba(0, 0, 0, .1);
-    }
-
-    @media screen and (max-width: 992px) {
-      div#navbarsExample09 {
-        order: 4;
-      }
-
-      div#settingIcon {
-        order: 3;
-
-      }
-    }
-
-    @media (max-width: 576px) {
-      main {
-        padding: 2rem !important;
-      }
-
-      a#logo {
-        font-size: 0.98rem;
-      }
-    }
-  </style>
-</head>
-
-<body>
-  <nav class="navbar navbar-expand-lg navbar-light bg-light rounded">
-    <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarsExample09"
-      aria-controls="navbarsExample09" aria-expanded="false" aria-label="Toggle navigation">
-      <img src="../images/menu.png" />
-    </button>
-
-    <a id="logo" class="navbar-brand text-success ml-auto mr-auto" href="#">
-      <img src="../images/logo.png" width="30" height="30" alt="">
-      EmissionPlanner
-    </a>
-
-
-    <div class="collapse navbar-collapse" id="navbarsExample09">
-      <ul class="navbar-nav">
-        <li class="nav-item active">
-          <a class="nav-link" href="#">Home <span class="sr-only">(current)</span></a>
-        </li>
-        <li class="nav-item">
-          <a class="nav-link" href="#">Map</a>
-        </li>
-        <li class="nav-item">
-          <a class="nav-link" href="#">About</a>
-        </li>
-        <li class="nav-item">
-          <a class="nav-link" href="#">Contact Us</a>
-        </li>
-      </ul>
-    </div>
-
-    <div id="settingIcon" class="dropdown">
-      <!-- <button class="btn " type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true"
-        aria-expanded="false">
-        <img src="../images/settingsIcon.png" />
-      </button>
-      <div class="dropdown-menu dropdown-menu-right" aria-labelledby="dropdownMenuButton">
-        <a class="dropdown-item" href="#">History</a>
-        <a class="dropdown-item" href="#">Setting</a>
-        <a class="dropdown-item" href="#">Sign Out</a>
-      </div> -->
-    </div>
-  </nav>
-
-  <main class="container d-lg-flex align-items-lg-center height-85vh--lg p-5">
-    <div class="space-2 space-0--lg mt-lg-8">
-      <div class="row justify-content-lg-between align-items-lg-center text-success">
-        <div class="col-lg-5 mb-7 mb-lg-0 ">
-          <!-- Title -->
-          <span class="d-block text-uppercase mb-2">Join Us</span>
-          <h1 class="display-4 font-size-48--md-down  mb-0">For the Change</h1>
-          <!-- End Title -->
-        </div>
-
-        <div id="planTrip" class="col-lg-5">
-          <!-- Signup Form -->
-          <div class="bg-white shadow-sm rounded p-3">
-            <form class="js-validate" novalidate="novalidate">
-
-              <!-- Input -->
-              <div class="js-form-message mb-3">
-                <div class="js-focus-state input-group input-group form">
-                  <input type="text" class="form-control form__input form-control-lg" name="username" required=""
-                    placeholder="From" aria-label="Enter your username">
-                  <button class="btn" type="button" style="margin-left: 10px; padding: 0"><img form-control
-                      src="../images/autolocate.png" alt=""></button>
-                </div>
-              </div>
-              <!-- End Input -->
-
-              <!-- Input -->
-              <div class="js-form-message mb-3">
-                <div class="js-focus-state input-group input-group form">
-                  <input type="email" class="form-control form-control-lg form__input" name="email" required=""
-                    placeholder="Destination" aria-label="Enter your email address">
-                </div>
-              </div>
-              <!-- End Input -->
-
-              <button type="submit" class="btn btn-lg btn-block btn-primary bg-success text-white mb-3">Plan Your Trip
-                Now</button>
-              <button type="submit" class="btn btn-lg btn-inline btn-primary bg-success text-white mr-1"
-                style="width: 48%">Sign In</button>
-              <button type="submit" class="btn btn-lg btn-inline btn-primary bg-success text-white float-right"
-                style="width: 48%">Sign Up</button>
-            </form>
-          </div>
-          <!-- End Signup Form -->
-        </div>
-      </div>
-    </div>
-  </main>
-
-  <footer class="footer-copyright text-right py-3" style="background-color: #85D139;">
-    <a class="mr-3" href="#" class=>FAQs</a>
-  </footer>
-
-  <!-- Firebase init -->
-  <script src="https://www.gstatic.com/firebasejs/5.5.8/firebase.js"></script>
-  <script src="https://www.gstatic.com/firebasejs/5.5.8/firebase-app.js"></script>
-  <script src="https://www.gstatic.com/firebasejs/5.5.8/firebase-auth.js"></script>
-  <script src="https://www.gstatic.com/firebasejs/5.5.8/firebase-database.js"></script>
-  <script src="https://cdn.firebase.com/libs/firebaseui/3.1.1/firebaseui.js"></script>
-
-  <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
-    integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
-    crossorigin="anonymous"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"
-    integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1"
-    crossorigin="anonymous"></script>
-  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"
-    integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM"
-    crossorigin="anonymous"></script>
-  <script src="../js/firebase_init.js"></script>
-  <script src="../js/index.js" type="text/javascript"></script>
-</body>
-
-</html>
