--- html/index.html
@@ -1,51 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <title>Emissions Planner</title>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link
-  rel="stylesheet"
-  href="https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
-  integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T"
-  crossorigin="anonymous"
-  />
-  <!-- The core Firebase JS SDK is always required and must be listed first -->
-<script src="https://www.gstatic.com/firebasejs/6.0.0/firebase-app.js"></script>
-
-<!-- TODO: Add SDKs for Firebase products that you want to use
-     https://firebase.google.com/docs/web/setup#config-web-app -->
-
-<script>
-  // Your web app's Firebase configuration
-  var firebaseConfig = {
-    apiKey: "AIzaSyDd7Y3dwiIxCrCDlID9NvpTXDCHhiPDb6M",
-    authDomain: "team-04-comp-2930.firebaseapp.com",
-    databaseURL: "https://team-04-comp-2930.firebaseio.com",
-    projectId: "team-04-comp-2930",
-    storageBucket: "team-04-comp-2930.appspot.com",
-    messagingSenderId: "432244259664",
-    appId: "1:432244259664:web:77b09583bf537e23"
-  };
-  // Initialize Firebase
-  firebase.initializeApp(firebaseConfig);
-</script>
-</head>
-
-<body>
-  <div id="root"></div>
-  <script src="https://unpkg.com/react/umd/react.development.js"></script>
-  <script src="https://unpkg.com/react-dom/umd/react-dom.development.js"></script>
-
-  <script
-    src="https://unpkg.com/react-bootstrap@next/dist/react-bootstrap.min.js"
-    crossorigin></script>
-
-<script>var Alert = ReactBootstrap.Alert;</script>
-<script src="https://unpkg.com/babel-standalone"></script>
-  <script src="../js/index.js" type="text/jsx"></script>
-
-
-
-</html>
