--- js/firebase_init.js
@@ -1,12 +1,0 @@
-// Your web app's Firebase configuration
-var firebaseConfig = {
-  apiKey: "AIzaSyDd7Y3dwiIxCrCDlID9NvpTXDCHhiPDb6M",
-  authDomain: "team-04-comp-2930.firebaseapp.com",
-  databaseURL: "https://team-04-comp-2930.firebaseio.com",
-  projectId: "team-04-comp-2930",
-  storageBucket: "team-04-comp-2930.appspot.com",
-  messagingSenderId: "432244259664",
-  appId: "1:432244259664:web:77b09583bf537e23"
-};
-// Initialize Firebase
-firebase.initializeApp(firebaseConfig);
