+++ public/README.md
@@ -1,0 +1,2 @@
+# EmissionsPlanner
+COMP 2930 Project Emissions Planner
