+++ public/html/addVehicle.html
@@ -1,0 +1,151 @@
+<!doctype html>
+<html lang="en">
+
+  <head>
+        <link rel="stylesheet" href="../css/addVehicle.css">
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
+      integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous" />
+    <link href="https://fonts.googleapis.com/css?family=Rubik:300,400,500" rel="stylesheet">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <!-- The core Firebase JS SDK is always required and must be listed first -->
+    <script src="https://www.gstatic.com/firebasejs/6.0.1/firebase-app.js"></script>
+
+    <!-- TODO: Add SDKs for Firebase products that you want to use
+        https://firebase.google.com/docs/web/setup#config-web-app -->
+    <script src="https://www.gstatic.com/firebasejs/6.0.1/firebase.js"></script>
+    <script>
+            // Your web app's Firebase configuration
+                var firebaseConfig = {
+                    apiKey: "AIzaSyDd7Y3dwiIxCrCDlID9NvpTXDCHhiPDb6M",
+                    authDomain: "team-04-comp-2930.firebaseapp.com",
+                    databaseURL: "https://team-04-comp-2930.firebaseio.com",
+                    projectId: "team-04-comp-2930",
+                    storageBucket: "team-04-comp-2930.appspot.com",
+                    messagingSenderId: "432244259664",
+                    appId: "1:432244259664:web:77b09583bf537e23"
+                };
+                // Initialize Firebase
+                firebase.initializeApp(firebaseConfig);
+            </script>
+    <title>Emission Planner</title>
+  </head>
+  
+	<body>
+            <div id="header">
+                    <nav class="navbar navbar-expand-lg navbar-light bg-light rounded">
+                        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarsExample09"
+                            aria-controls="navbarsExample09" aria-expanded="false" aria-label="Toggle navigation">
+                            <span class="navbar-toggler-icon"></span>
+                        </button>
+                    
+                        <a class="navbar-brand text-success" href="#">
+                            <img src="../images/logo.png" width="30" height="30" alt="">
+                            EmissionPlanner
+                        </a>
+                    
+                        <a href="#">
+                            <img class="mr-2" src="../images/avatar.png" width="30" height="30" alt="">
+                            Sign In
+                        </a>
+                    
+                        <div class="collapse navbar-collapse" id="navbarsExample09">
+                            <ul class="navbar-nav mr-auto">
+                            <li class="nav-item">
+                                <a class="nav-link" href="#">Home <span class="sr-only">(current)</span></a>
+                            </li>
+                            <li class="nav-item">
+                                <a class="nav-link" href="#">Map</a>
+                            </li>
+                            <li class="nav-item">
+                                <a class="nav-link" href="#">About</a>
+                            </li>
+                            <li class="nav-item">
+                                <a class="nav-link" href="#">Contact Us</a>
+                            </li>
+                            </ul>
+                        </div>
+                    </nav>
+                </div> 
+		<div>
+
+                <div class="tabs">
+                    <nav class="tabs-nav">
+                        <ul class="tabs-nav-items">
+                            <li class="tabs-nav-item active-tab-nav-item">
+                                <a href="#section-1" class="tabs-nav-btn">Add Vehicle</a>
+                            </li>
+                            <li class="tabs-nav-item">
+                                <a href="#section-2" class="tabs-nav-btn">Your Vehicles</a>
+                            </li>
+                            
+                        </ul>
+                    </nav>
+
+                    <div class="tabs-content">
+                        <div id="first">
+                        <section id="section-1" class="tabs-content-item active-tab-content-item">
+                           
+                            <div id="year-label">
+                                    <label for="year" >Year</label>
+                                   
+                            </div>
+                            <div>
+                                    <select class="year-box" id="box1" onchange="selectYear()">
+                                            <option disabled selected value> -- select an option -- </option>
+                                          </select>
+                            </div>
+
+                            <div id="year-label2">
+                                    <label for="year" >Make</label>
+                                   
+                            </div>
+                            <div>
+                                    <select class="year-box" id="box2" onchange="selectMake()">
+                                            <option disabled selected value> -- select an option -- </option>
+                                            
+                                          </select>
+                            </div>
+
+                            <div id="year-label2">
+                                    <label for="year" >Model</label>
+                                   
+                            </div>
+                            <div>
+                                    <select class="year-box" id="box3">
+                                            <option disabled selected value> -- select an option -- </option>
+                                            
+                                          </select>
+                                          <button onclick="newPerson()" class="button" style="vertical-align:middle"><span>ADD </span></button>
+                            </div>
+
+                        </section>
+                        </div>
+                        <section id="section-2" class="tabs-content-item">
+                            <div>
+                                <div id="why">
+                                        
+                                </div>
+                            </div>
+                        </section>
+
+                    </div>
+
+                    <footer>
+
+                            <ul class="footer-nav">
+                    
+                              <li class="full"><a href="#portfolio">About</a></li>
+                              <li class="full"><a href="#contact">FAQs</a></li>
+                            </ul>
+                          </footer>
+                          
+                </div>
+                <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
+    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
+                <script type="text/javascript" src="http://ajax.googleapis.com/ajax/libs/jquery/1.5/jquery.min.js"></script>
+		<script src="../js/addVehicle.js"></script>
+  </body>
+  </html>
