+++ public/js/logmytrip.js
